requirements.txt
black==25.9.0
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
kiwisolver==1.4.9
matplotlib==3.10.6
mypy_extensions==1.1.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytokens==0.1.10
pytz==2025.2
ruff==0.13.2
seaborn==0.13.2
six==1.17.0
tzdata==2025.2
reportlab==4.2.5
python-pptx==0.6.23
jupyter==1.1.1
ipykernel==6.29.5
notebook==7.2.2
